fix(Dial): 调整拨弹电机单发角度定义和步兵拨弹速度限制
调整英雄机器人拨弹电机单发角度定义，从角度值改为编码器值。
降低步兵机器人拨弹电机速度限制，从4000降到2000。
更新相关日志文件时间戳。

+++ Hero_board1_SJW - Hero_2_Upstair/BSP/BSP.h/BSP_BoardCommunication.h
@@ -50,9 +50,9 @@ typedef struct
 	uint8_t Deployment_Flag;			// 部署模式开关
 	uint16_t Blood_Volume;				// 机器人当前血量
 	uint16_t game_start;				// 比赛开始标志（未开始为1，准备、自检、比赛中为1）
-	int8_t F_Track_Angle;				// 前履带角度
-	int8_t R_Track_Angle;				// 后履带角度
-} ControlMessge;	
+	int16_t F_Track_Angle;				// 前履带角度
+	int16_t R_Track_Angle;				// 后履带角度
+} ControlMessge;
 
 /* 图传遥控客户端下发键鼠遥控信息:0x0304  Byte 12*/
